changed int to bigint for forbidden vertices

+++ src/contraction/src/contractGraph.c
@@ -83,7 +83,7 @@ process(char* edges_sql,
     //PGR_DBG("Load data");
 #ifdef DEBUG
     pgr_edge_t *edges = NULL;
-    int64_t total_tuples = 0;
+    size_t total_tuples = 0;
     if (num_cycles < 1) {
         PGR_DBG("Required: atleast one cycle\n");
         (*result_count) = 0;
@@ -195,7 +195,7 @@ contractGraph(PG_FUNCTION_ARGS) {
         /*                                                                             */
         /*******************************************************************************/
 
-        funcctx->max_calls = result_count;
+        funcctx->max_calls = (uint32_t)result_count;
         funcctx->user_fctx = result_tuples;
         if (get_call_result_type(fcinfo, NULL, &tuple_desc) != TYPEFUNC_COMPOSITE)
             ereport(ERROR,
